requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy==2.1.2
torch==2.7.1+cpu
torchvision==0.22.1+cpu
pillow==10.4.0
tqdm==4.66.4
pyyaml==6.0.2
pytest==8.3.3
